requirements.txt
pyhwk @ git+https://github.com/clusterhack/pyhwk@main
#pytest  # Future use
#pandas
#matplotlib
#jupyter
#ipywidgets
#nbclient
